Add Travis CI build.

+++ update-version.sh
@@ -1,11 +1,17 @@
 #!/bin/bash
 
-current_version="$(git describe)"
-cached_version="$(cat version.txt)"
+current_version="$(git describe --always)"
 
-if [ "$current_version" != "$cached_version" ]; then
-	echo "'$current_version' != '$cached_version'"
-	echo "$current_version" > version.txt
+if [ -f version.txt ] ; then
+	cached_version="$(cat version.txt)"
+
+	if [ "$current_version" != "$cached_version" ]; then
+		echo "'$current_version' != '$cached_version'"
+		echo "$current_version" > version.txt
+	else
+		echo "up to date at $current_version"
+	fi
 else
-	echo "up to date at $current_version"
+	echo "version.txt not found"
+	echo "$current_version" > version.txt
 fi
